requirements.txt
django==2.1.2
gunicorn==19.9.0
psycopg2==2.7.6.1
whitenoise==3.3.1
django-allauth==0.38.0
django-environ==0.4.5
django-extensions==2.1.9
django-secure==1.0.1
djangorestframework==3.9.0
coreapi==2.3.3
mysqlclient==1.4.2.post1
django-mysql==3.2.0